Fixes typo in import and renames corresponding module
